Making it installable
